compressors: mksquashfs/unsquashfs broken if GZIP_SUPPORT is not defined
Fix mistake in gzip_comp_ops structure used when no GZIP_SUPPORT
is defined.  The missing NULLs in the structure caused
mksquashfs/unsquashfs to believe no compressors were present.

While fixing this also make the long sequence of NULLs unnecessary
by rearranging the structure.  This ensures that this bug should
not arise in the future as the unsupported compressor structures now
don't need to be updated every time the number of compressor
functions are changed.

Signed-off-by: Phillip Lougher <phillip@lougher.demon.co.uk>