mksquashfs: move fragment index allocation to fragment creation
Move allocation of the fragment index to when a new empty fragment
is obtained, rather than when the filled fragment is queued to the
fragment_deflate threads.

Doing this allows multiple fragments to be open, and being filled at
the same time, the choice of which fragment is to be used for a
particular file-tail depending on various attributes.

Signed-off-by: Phillip Lougher <phillip@lougher.demon.co.uk>