mksquashfs: add exclude action
Add an additional action called exclude, which takes no parameters.
Any files which the expression returns TRUE will be excluded.

Curently, the only test operator is name(wildcard), which matches
on the file name (leaf name without preceeding pathname).  So, currently
the functionality is equivalent to non-anchored excludes i.e.

-action exclude=name(*.o)

is equivalent to

-e "... *.o".

Do we really want yet another exclude facility?   Well no if it
remains equivalent to just anchored excludes, but once all the
anticipated test operators are implemented then it becomes much more
powerful than the existing functionality.  Want to exclude based
on file age, size or ownership etc.?  That will be possible soon.

The cool thing is the action infrastructure is shared with all other actions,
and the exclude infrastructure is already there and "comes free", and so
the exclude action code is minimal about 20 lines!

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>