mksquasfs: handle empty files specially in deflator()
Add special case for empty files in the deflator thread, without
this they're handled differently depending on whether the sparse
files flag is set, the fragment flag is set, or whether they fall
through to be considered as a block to be compressed by the deflator
thread.  In all cases the end result is the same (an empty file is
handled specially in the write_file() routine irrespective of
whether it has been handled as a sparse file of zero bytes, a
fragment of zero bytes, or a data block of zero bytes in the
deflator thread), but it is messy that they're handled differently
depending on filesystem flags.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>