actions: parse all the action arguments present rather than the expected number
Parse all the arguments present in the action statement, and
then compare against the expected number of arguments, rather than
parse only the expected number of arguments, and fail as soon as
it is known there's too little or too many arguments present.

This change enables actions with variable length arguments to be handled.
An action with a "vararg" list has an expected number of arguments of -2.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>