mksquashfs: optimise dir_ent structure by removing pathname member
All dir_ent structures stored both the name and pathname
of each file scanned from the source directories.  This was
rather wasteful of memory, as the majority of pathnames could
be computed from the parent directory's pathname and the
dir_ent's name.

Exceptions to this are names which have had a _1, _2 etc. appended due
to name clashes, pseudo files and cases where multiple directories
and files have been specified on the mksquashfs command line, in which
case the root directory will likely consist of entries each with a
different pathname.  These cases have to be handled specially, either
with the new source_name member holding the original name, or in the
cases where there's a completely different pathname, by the use of the
nonstandard_pathname member.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>