action: fix subpathname to work as expected
Previously subpathname relied on not using FNM_PATHNAME in fnmatch to
obtain a subpathname match by utilising the behaviour that "*" could match
"/" and so ensuring a pathname pattern would match pathnames "underneath"
the directories directly matched by the pathname pattern.

i.e. a pattern match of "a/b/*" will match "a/b/c", but it will also
match "a/b/c/d", the "*" matching the "c/d" component of the pathname.

This provided the intended semantics of "subpathname", that any
file/directory which is a subpath of the pattern will match.
This fulfils the requirement of various actions where we often
only want the action to have effect on files/directories contained within
a particular directory (or directories if wildcards have been used
in the pattern match).

It, however, has an unwanted and unexpected side-effect, given the
pathname pattern "a/*/c/*" the user might expect it to only
match on directories called "c" which are children of *one* unknown
directory which is a child of directory "a", but in fact because "*"
matches on "/" too, the pathname "a/b/extra-directory/c/filename"
will match.

This is considered broken behaviour, and so subpathname has been
reimplemented to walk the pattern match, and to only apply the
patten match to the appropriate leading part of pathname, ensuring
pathnames which are subpaths match while allowing FNM_PATHNAME to be
used in fnmatch ensuring the unexpected behaviour detailed above does not
occur.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>