mksquashfs: Remove qsort and add a bottom up linked list merge sort
Now that directory entries are stored in a linked list, rather
than an array, move to using a bottom up linked list merge sort.

Qsort and other O(n log n) algorithms work well with arrays but not
linked lists.  Merge sort another O(n log n) sort algorithm on the other
hand is not ideal for arrays (as it needs an additional n storage
locations as sorting is not done in place), but it is ideal for linked
lists because it doesn't require any extra storage.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>