actions: fix root parent inode computation and action depth() test
The previous commit added the ability to execute actions on the root
inode ("/"), however, it introduces a couple of subtle bugs.

These bugs are caused by the change to create_dir_entry() where
the root dir_ent's parent directory is changed from NULL to the root
directory, creating the situation where the parent of the root directory
is itself.  This change was ultimately because the action code relies
on the parent directory not being NULL (previously the action code was
not executed on the root dir_ent and so the parent was never NULL).

This, however, introduces the following two subtle bugs:

1. Root parent inode number computation is broken

get_parent_no() returns the parent inode number for a directory when
this directory inode is being added to the inode table.  get_parent_no()
relies on the special case that the parent is NULL to detect when it is
dealing with the root directory inode.  As the root parent is now
itself, we effectively end up with ".." and "." in the root directory
having the same inode number, which is obviously wrong.

Due to the way mount works, the parent directory ("..") of a mounted
filesystem is the parent of the mount point (as you would expect),
and so this bug is largely hidden, however, the readir() system
call directly queries the filesystem, and this does return ".." and
"." as having the same inode number

% myls .

. 961
.. 961
.git 1
0001-Apply-actions-to-the-root-directory.patch 826

2. The depth() operation tests the depth of a particular file or
directory, files in the root directory have depth 1, files in immediate
subdirectories have depth 2 and so on.  Making the parent of the root
dir_ent be the root introduces the suble bug that the root inode
will be incorrectly considered to be at depth 1, when it should be at
depth 0.

The correct solution is to add a dummy parent dir_info at depth 0,
and to check for that in get_parent_no().

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>