actions: add move() action
Add move action allowing files and directories to be moved
and/or renamed.

Syntax -action "move(DEST)@<test_ops>"

Semantics are the same as Unix mv command.

Rename SOURCE to DEST (if it doesn't exist) or
Move SOURCE to DEST (if it exists and is a directory)

SOURCE is any file matching the test_ops expression.

For instance

-action "move(/source)@name(*.[ch])"

move all *.[ch] files to the directory "/source"

More complex combinations of test_ops are supported by default.

For instance

Move all files not owned by monitor smaller than 1024 bytes to "/smallfiles"
but ignore any files matching "*.[ch]"

-action "move(/smallfiles)@ !uid(monitor) && size(-1024) && !name(*.[ch])"

Multiple moves are evaluated "simultaneously", so the results of one
move do not affect other moves.

This is to prevent unanticipated side-effect.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>