mksquashfs: fix stack overflow in write_fragment_table()
Elso Andras has reported a SegV in mksquashfs squashing
a 100 GB filesystem with 5 million files.

The SegV turns out to be a stack overflow in write_fragment_table()
where it is allocating a temporary copy of the fragment_table
on the stack, this is generally not a good idea as the
fragment table can be very large.

This bug is easily fixed.  Since moving to a fixed little endian
layout without bitfields, allocating a temporary copy is no longer
necessary because any swapping (mksquashfs running on a
big-endian architecture) can now be done inplace.  This code should
have been updated to do this back in 2009, but the code somehow
didn't get updated.

Thanks to Elso Andras for reporting this bug, and for unusually
providing sufficient debug information in the bug report that
it could be tracked down without any folllow up emails!

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>