unsquashfs: fix oversight in directory permission setting
When descending the directory structure Unsquashfs used to
create directories with the permissions obtained from
the filesystem and then unsquash the contents of
the directory, descending into sub-directories as appropriate.

This has in hindsight an obvious oversight.  Directories in
Squashed filesystems may contain files and directories but
may be read-only and even lack the execute bit.  Setting
such permissions means creating the directory contents
fails because the directory is created without the
necessary permissions.

Fix this by initially creating directories with User
read/read/execute, and only set the permissions when
unwinding after the directory contents has been produced.

This fix is slightly complicated by the parallel and
asynchronous nature of file writing.  As file writing
is done in parallel in another thread, when the main
thread unwinds the file writes may still be queued,
leading to the case that the writer thread will fail
to set the attributes for the file if any of the
parent directories in the pathname lack the execute
bit (lack of the write bit is OK as we don't need to write
to the directory, but we need the execute bit to be set
to descend into it).

Guard against this from happening by having the main thread
queue the directory set attributes operation to the writer
thread.  This ensures all relevant file write operations are
completed before the directory permissions are changed.

Thanks to Andy Lutomirski for reporting this bug.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>