sort: Rewrite sort file processing
The previous version suffered from a number of nasty problems.

1. It couldn't handle filenames with spaces.

2. Short lines would cause the following line to be read, in the
   unlikely event the following filename was a number this would
   lead to incorrect parsing.   Mostly such badly formed files would
   lead to failures later on, but, such behaviour is incorrect and
   should not be relied upon.

3. Lines with trailing junk after the priority would parse the junk
   as further filename priority pairs, potentially leading to incorrect
   parsing.  Again, mostly such badly formed files would
   lead to failures later on, but, such behaviour is incorrect and
   should not be relied upon.

4. If a line exceeded the line limit (16384 bytes), the line would be
   silently truncated.

5. If file reading failed for any other reason than EOF this was
   not reported.

Fix these issues, and also add support for comment lines
(prefixed by #).

Files with spaces or starting with # can be escaped by \, i.e.

\#filename 1234
hello\ world 1234

Note the tightening up of parsing may cause badly formatted files
which previously worked to fail, this is regretted, but,
it is considered that the benefits of preventing badly formatted files
to cause unexpected behaviour in mksquashfs outweighs this.

This is mainly a consequence of two CVEs being raised against
squashfs-tools where they did not properly check user input,
leading to vulnerabilities.  As a result of this, all user input
processing has been tightened up.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>