mksquashfs: flag lseek error in writer as probable out of space
If the writer thread hits "lseek error" it is probably because the
destination file is out of space, this is especially true if the
destination is a block device, but could conceivably happen on some
(non Linux?) filesystems too if they don't implement holes.  This
is because lseeking to the write position could be beyond the end
of the available destination file end, and hence fail before
attempting to write.

Recent correspondence with users shows that "Lseek error" is
meaningless, and so augment the error message with an "Out of
Space" error which is presumbly something they will understand.

Note we can't use BAD_ERROR here because it will cause mksquashfs to
immediately exit, and we instead want to signal the the main thread to
exit.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>