read_xattrs: Use error.h for read_xattrs as well
and get rid of the separate error macro implementations
in read_xattrs.  This finally unifies all the error handling
into error.h

Read_xattrs.c has for a long time been an annoying special
case because it is used by both Mksquashfs and Unsquashfs,
it couldn't use the error.h infrastructure used by Mksquashfs
because that pulled in dependencies on Mksquashfs that don't
exist in Unsquashs, and it couldn't use the unsquashfs.h
infrastructure used by Unsquashfs because that pulled in
dependencies on Unsquashfs that don't exist in Mksquashfs,
and so it had to define its own macros.  But this has meant
it has not benefited from the improvements in error handling
added to the Mksquashfs/Unsquashfs error infrastructure.

The previous commits have finally broken the conflicting
dependencies introduced by error.h, unnecessary dependencies
have been removed and both Mksquashfs and Unsquashfs now
provide the necessary functions used by error.h...

So we can finally drop the read_xattrs.c only error macro
implemenation :-)

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>