mksquashfs: get rid of update_progress_bar()
Don't display the progress bar every time a file is written
as the background 1/4 second display is sufficient.

Removing this eliminates an unanticipated main thread dependency
on the progress bar thread.  By grabbing the progress_mutex mutex
before doing the pthread_cond_signal() to wake up the progress bar
thread it means the main thread waits on the progress_mutex mutex
if the progress bar thread already has it grabbed.  Normally
when the progress bar wakes up it grabs the progress_mutex mutex
prints the progress bar and quickly releases the mutex, and so
the main thread will never wait on the mutex.  However, if
the progress bar printing is slow, the main thread can wake up
the progress bar thread when a file is written, and it can
still be printing the progress bar when the main thread writes
the next file, causing the main thread to wait.

In effect this throttles file writing to the speed of the progress
bar printing, normally this has no effect because progress bar
printing is faster, however, if the output terminal is running
over a slow link this can become a bottleneck.  The author
has noticed this running Mksquashfs on a VT102 connected via a
9600 baud serial line.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>