mksquashfs: print file being squashed when ^\ (SIGQUIT) typed
If ^\ is typed on the keyboard whilst Mksquashfs is squashing files
(progress bar being printed), then display the filename of the
file being squashed.

The progress bar is useful to display how far blockwise/percentage-wise
Mksquashfs has got, but, it doesn't display filenames, and it is
often nice to get a snapshot filewise where Mksquashfs has got to.
Obviously, you can use -info for this, but this is a "heavy-weight"
option which displays every filename, it not only leads to
screenfulls of text, but it also can slow Mksquashfs down as it
is bottlenecked by all the filename printing.

Adding the ^\ option offers a best of both worlds, Mksquashfs
displays the progress bar which gives nice feedback on the
progress of Mksquashfs but in a way which does not generate
lots of text and doesn't slow Mksquashfs down, and typing ^\
gives you the filename of file being squashed but only when
desired, which doesn't generate screenfulls of text or slow
Mksquashfs down.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>