frag_deflate_thread: Move mutex unlock back to before queue put
Originally the mutex unlock was before the queue_put to
the writer thread.  This ensured in rare cases where the
queue was full (writer thread cannot keep up), the thread
would not go to sleep holding the mutex, making other threads
block waiting on the mutex (and in-directly on the write thread).

This was changed when adding the pthread_cancel_{push/pop} calls.

Restore original behaviour.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>