info: dump queue and cache status if ^\ hit twice within one second
Previously the filename being Squashed would be printed on hitting
^\, and the queue and cache status would be printed if
SIGHUP was received.

However, sending SIGHUP to generate a queue and cache status dump
is a PITA, it cannot be generated from the keyboard and must be sent
from another terminal window.

Fix this by overloading ^\.  Now if ^\ is hit twice within one second
dump the queue and cache status.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>