mksquashfs: rename from_reader queue to to_deflate
The from_reader queue used to collect all the data queued from
the reader thread (hence its name).  Data from the reader thread
is now queued to the to_main queue (uncompressed fragments etc.) and
to the from_reader queue, which is now soley used to queue
blocks to the deflator thread(s).  So rename to to_deflate which
now more accurately reflects how it is used.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>