xz_wrapper: fix sizeof() in xz_display_options()
Sizeof() was incorrectly sizeof(comp_opts) or size of pointer,
rather than the size of the structure.  On 64-bit systems it
worked because it so happens that size pointer == 8, which is the
size of the structure.

This bug has shown up now on a 32-bit system (Raspberry Pi), where
the size of pointer is 4 bytes.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>