unsquashfs: fix cache used count statistic
Fix cache used count statistic (only used in statistics dumping when
^\ is hit twice).

Occasionally when dumping the statistics I noticed the cache used
count was negative.  This proved to be because I forgot to
increment the cache used count in cache lookup when the block
was found unused on the free list.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>