mksquashfs/unsquashfs: fix mtime signedness
Since time-immemorial Squashfs has treated mtime as unsigned.
This is a bug, mtime is signed, and dates previous to the
epoch of 1 Jan 1970 should be treated negatively.

This bug does not affect mksquashfs, as it stores the mtime
in the filesystem correctly.  Unsquashfs, however, treats
mtime as unsigned, and it will report (-lls) and
set any file with a timestamp before the epoch incorrectly.

This is easily fixed by changing the definition of mtime
(and mkfs_time) to be "int" rather than "unsigned int".
Because the signedness is changing, the swap code also
has to be undated.

Reported-by: socketpair@gmail.com
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>