unsquashfs: fix mtime signedness for legacy layouts
Since time-immemorial Squashfs has treated mtime as unsigned.
This is a bug, mtime is signed, and dates previous to the
epoch of 1 Jan 1970 should be treated negatively.

This bug has never affected Mksquashfs, and so there is a small but
unlikely possibility that there can exist legacy Squashfs filesystems
with timestamps prior to the epoch.

Due to this, This commit fixes mtime signedness for the legacy
1.x, 2.x and 3.x layouts, so negative mtime is correctly handled by
Unsquashfs.

Note the legacy swapping code does not have to be updated
because it never interprets signedness of the data.

signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>