Mksquashfs: split checksum_flag
Split checksum_flag in file_info structure into two, one
each for fragments and blocks.

This allows the pre frag duplicate check routine to skip the
block checksum computation, and in future it will allow the
fragment checksum for all files in a fragment (when the fragment
is read off disk in appending) to be computed, skipping the
block checksum computation.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>