Mksquashfs: introduce additional per CPU fragment process threads
The following commit

	commit 287f3ef6b1cd670f3ff347dcaaa7b0bade03096b
	Date:   Sun May 19 04:27:38 2013 +0100
	mksquashfs: queue fragment and empty file buffers directly
		to main thread

moved fragment and empty file buffer processing away from the
deflate thread(s), and queued them directly to the main thread.
This was done to ensure the "queue dump" facility added to
Mksquashfs when it showed the "reader -> deflate queue size"
was showing the number of *file buffers* waiting to be
compressed by the deflate thread(s).  Previously when the
queue also held fragments (because sparse checking of fragments
was piggy-backed onto the deflate threads), the queue size shown
included both fragments and file buffers waiting to be compressed.
This limited the usefulness of the queue dump in determining what
was happening within Mksquashfs.

A consequence of this, however, was that sparse checking of
fragments was moved to the main thread.  In hindsight,
this could cause a performance regression in certain cases because
there is only one main thread (it is not per CPU) and it is tasked
with overall coordination of Mksquashfs, adding the additional
burden of sparse checking could bottleneck the main thread,
leading to bottlenecking of the other threads and a loss of
parallelism.

In general this change broke the general rule that the main thread
should only deal with coordination and processing should be
done on other per CPU threads.

So fix this by introducing an additional per CPU fragment processing
thread which deals with sparse checking of fragments.

Note: it is expected that moving sparse checking to the main thread may
only have caused a problem with duplicate checking where the main thread
becomes otherwise engaged, and even then only in the extremely rare
case many fragments are sparse.  In general, fragments are mostly
trivially verifiable as non-sparse (first few bytes non-zero), and the
sparse check is therefore extremely fast.

Moving sparse checking to fragment processing threads is a first stage
iteration, the aim is to move the other more CPU intensive fragment
processing which is currently done on the main thread to the new
fragment processing threads too, as this should improve performance.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>