caches-queues-lists: Rearrange struct file_buffer
Move long longs to the start of the structure, on 32-bit
nachines these are, obviously, longer than pointers.
Making them first avoids the necessity of the compiler to insert
4-bytes of padding between the pointers and the long longs to get
8 byte alignment.

The reduction of the structure size from 68 bytes to 64 bytes,
also means the compiler doesn't have to pad the structure out to
72 bytes to get an 8 byte multiple (again to ensure 8 byte
alignment of long longs).

This makes a saving of 8 bytes on 32-bit machines.  On
64-bit machines there is no difference.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>