caches-queues-lists: Merge sequence and index in struct file_buffer
Now that the deflator thread(s) do not hash the writer block, the index
and sequence fields are never in use at the same time.

The sequence field is needed to correctly order the blocks output by the
seq_queue, but at this time, none of the blocks are hashed and index is
unused.  Once the blocks are obtained from the seq_queue they are either
discarded (uncompressed fragment blocks and empty file blocks) and never
hashed, or they are hashed by the main thread later.  But when this happens
the sequence field is no longer in use.

So we can merge the two fields which saves 8 bytes.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>