mksquashfs: add a new -mem <mbytes> option
Add a new -mem <mbytes> option, that sets the amount of memory
Mksquashfs uses to <mbytes>.  Internally, the <mbytes> is distributed
amongst the four caches: reader cache, fragment cache, block writer cache,
and fragment writer cache in an optimal ratio, as is already done
when deciding how to allocate the default physical memory amongst the
caches.

This option is intended to be much more understandable than the
previous -read-queue, -fragment-queue and -write-queue options, such
that people might actually use this option (whereas it is clear
nobody used the previous options, as they were evidently too difficult
to understand).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>