mksquashfs: undocument the -xxx-queue options in the usage text
Undocument the old deprecated -read-queue, -fragment-queue and
-write-queue options in the usage text.

The rationale for this is quite simple: we want to present a
single -mem option that is simple and obvious enough that people
will use it.

Rataining the -xxx-queue options adjacent to the -mem option will
solve nothing, we merely add to the set of options that in
conbination are so confusing.  The effectiveness of the -mem
option will be completely lost if people are confused how it
relates to the -xxx-queue options, should both be set etc.?

In effect the simplicty and obviousness of the -mem option only
is achieved if it is the only documented memory setting option.
To have it as just one option amongst many options will merely
add to confusion rather than reduce it.

But, because existing scripts may use the -xxx-queue options
"invisibly" retain them for backwards compatibilty, but don't
document then for future use.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>