mksquashfs: allow [KM]bytes to be specified as well as just KM
When an option takes a number with an optional multiplier, extend
the code to allow a trailing "bytes" to be specified.

Even though this is not going to be documented, people don't
read the manual text or usage text, and will specify [KM]bytes
and wonder why it's not allowed.

In otherwords, it's just too much potential pain not to allow
for a trailing bytes string.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>