gzip-wrapper: Add support for different compression strategies with -Xstrategy
Last minute improvement before release.  Add support to allow users to
try different compression strategies with zlib, not just the
Z_DEFAULT_STRATEGY.

	  -Xstrategy strategy1,strategy2,...,strategyN
		Compress using strategy1,strategy2,...,strategyN in turn
		and choose the best compression.
		Available strategies: default, filtered, huffman_only,
		run_length_encoded and fixed

I originally didn't implement this because I considered it too esoteric
to waste scarce development time on.  What's the point in developing
something if no one is going to use it?

But, on reviewing open tickets on Sourceforge I discover someone asked
for this back in 2011, and I'd forgotten about it in the meantime.  It's a
nice feature and if it looks like it might be used ...

So, here it is, a day's "heavy" coding and a delay in the release.  But,
it's probably worth it.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>