mksquashfs: fix abort on failure to get physical memory due to /proc missing
sysconf(_SC_PHYS_PAGES) which Mksquashfs uses to get the amount of
physical pages fails if /proc isn't mounted, returning -1.  This causes
Mksquashfs to abort with a physical memory less than SQUASHFS_LOWMEM
error.

Fix this by defaulting to the minimal value of 16 Mbytes, and
allow the user to use -mem to set the value to a higher value.  Also
issue a warning to this effect.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>