action: add eval() test operation
Follow path (arg1) and evaluate the expression (arg2)
in the context of the file discovered.  All attributes are updated
to refer to the file that is pointed to.

This test operation allows you to add additional context to the
evaluation of the file being scanned, such as "if current file is
XXX and the parent is YYY, then ..."  Often times you need or
want to test a combination of file status

If the file referenced by the path does not exist in
the output filesystem, or some other failure is experienced in
walking the path (see follow_path above), then FALSE is returned.

If you wish to evaluate the inode attributes of files which
exist in the source filestem (but not in the output filesystem then
use stat instead (see above).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>