mksquashfs: fix usage of "_" in new options
Gah, I like using "_" as the separator between words in
options.  But, a long time ago I noticed I'd ended up with "-"
as the separator in the majority case, and so I'm stuck with
"-" as the separator.

Fix verbose_option, true_option, false_option, action_file
and verbse_action_file.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>