action: change expression logging to use atom->args
Previously it used atom->test->args which is the arg count in the "generic"
test function.

Now that test function parsing has been improved to add variable arguments,
we should be using the new atom->args field which is the number of actual
arguments in the particular parsed test - i.e. the instance rather than
the definition.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>