actions: add new noop action
This action does exactly what it says, i.e. nothing.

It's designed to be used in conjunction with the verbose action
options to enable debugging of the test expression.

In otherwords, it allows you to check which files match on
your test expression but without doing anything to the filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>