mksquashfs/unsquashfs: use sys/sysmacros.h for makedev()
Newer versions of libc6 warn about the use of makedev() when it
is not included via <sys/sysmacros.h>. The full warning is:

```
warning: In the GNU C Library, "makedev" is defined
 by <sys/sysmacros.h>. For historical compatibility, it is
 currently defined by <sys/types.h> as well, but we plan to
 remove this soon. To use "makedev", include <sys/sysmacros.h>
 directly. If you did not intend to use a system-defined macro
 "makedev", you should undefine it after including <sys/types.h>.
```

This patch adds this include. I checked and sysmacros.h is available
at least until libc6 from 2012 (the oldest I had available) so the
risk of breaking builds is small.