unsquashfs: return exit 10 if writing fails for some reason
When using unsquashfs in scripts it is desirable to know if the
unsquashfs was successful or not. Currently only a limited set
of errors are reported via the exit code (e.g. unsupported
compression or invalid superblocks). However it would be great
if failures to write to the target directory (e.g. because of
out-of-space or permission issues) are reported.

This patch is doing that in a somewhat naive approach. I'm happy
to rework the patch. One option could be to reuse BAD_ERROR() in
unsquashfs and stop extracting as soon as a write error happens.
Another one might be to have FAIL_ERROR() that makes the
"FAILED = 1" part of the macro. I'm open for suggestions.

Also if there are concerns about backward compatibility I can
put this beinde a "-strict" (or similar) commandline flag.

This should also fix:
 https://github.com/plougher/squashfs-tools/issues/1