squashfs-tools: free old pointers if realloc() fails
There are several places where, when calling realloc() a reference to the
old pointer is kept in order to compute an offset in the new pointer (in
case it moved).

For those cases we cannot directly use the REALLOC_OR_ABORT() macro, so
might as well free the old pointer on the error path.
This makes sure cppcheck won't complain, and it's the correct thing to do.

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>