mksquashfs: Fix and improve -offset option
Previously the code used atol() to read the number.
But this doesn't do error checking or overflow checking.

Use the preexisting parse_numberll() instead.
This also adds the ability to use a K, M or G
suffix as well.

Most importantly, add a missing lseek offset
in process_fragments.c.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>