caches-queues-lists: Make index and sequence separate
These variables were merged into a union in 2014, because
it was known buffers held in the "seq queue" were not
indexed, therefore saving 8 bytes in the in-memory
file_buffer structure.

But we're going to be using a "seq queue" for indexed
buffers, and so these now need to be separate again.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>