mksquashfs: Add new -root-mode <mode> option
The impetus for adding yet another build tweak option, is
a bug report raised on Github back in 2017.

Simply put, when Mksquashfs is run with multiple sources on
the command line, it constructs a root directory containing
all the sources listed.  So far so good.

But, this containing root directory is conjured out of thin-air,
it is not read from the source(s).  So what permissions and
ownership should this root directory have?  I choose to have the
directory owned and group owned by the user running Mksquashfs, with
the permissions set to octal 0777.

This is ancient behaviour, which dates back to even before Squashfs 1.0
was released in October 2002.

Fast forward to this bug report in 2017.  It appears the reporter
is constructing Mksquashfs images using multiple sources on the
command line, and thus getting a root directory constructed above.

Their installer is then copying the filesystem, and getting a
root filesystem that is by default (due to the above 0777 permissions),
writable by anyone.  They would like the permissions to be set to
0755.

Now, it can be argued that is not a fault of Squashfs.  If they
want 0755 permissions, then it should not be beyond the wit of man
to alter them later.

But there again setting to 0777 was a completely arbitary choice on
my behalf many years ago.  I see no point in trying to argue this
is the correct setting.

But I can't change the default behaviour of Mksquashfs now, given
it's been like that way for 17 years since the beginning, and who
knows how many things rely on that behaviour.

So perhaps the obvious way forward is to add another build option,
which allows people to change the default permissions if necesssary.

As implemented, the new option allows the permissions on the root
directory to be changed for all scenarios, irrespective of the
origin of the permissions, whether 0777 in the above case or
read from the filesystem (in the case where a single directory
is specified on the Mksquashfs command line, and where the permissions
of the root directory is copied from that source directory).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>