unsquash-4: get fs table start and end points
To sanity check across the super-block we need to have
the start and end points of each filesystem table together
in one place.

For the id lookup, and export tables this is a little
bit tricky.  This is because they are in fact multi-level
tables.  The super-block points to a index table.  This
index table consists of pointers to compressed metadata
blocks.

The compressed metadata blocks are stored before the
index table (pointed to by the super-block), and
so the true start of the table is given by index[0] -
the first compressed metadata block.

For xattrs things are more complex, but, it is the
same idea.

We have to read at least part of the table before we
know where the start is - and by definition the end of
the preceeding table.

Unsquashfs doesn't need the exports table (if it exists),
and so we never bothered to read it.

But we now need to partially read it in order to work out
where the preceeding table ends.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>