mksquashfs: fix unaligned accesses in all_zero
struct file_buffer::data isn't guaranteed to be properly aligned to be
used as array of long. According to the C standard it is undefined what
happens. On most architectures this just results in ineffective code
that does the expected, however. On sparc64 this traps with a Bus error.

Bug-Debian: https://bugs.debian.org/932787