read_fs.c: Fix double free of inode_table on error
Static analysis shows a double free error if scan_inode_table()
returns with an error.

Fix this by changing the calling semantics so inode_table is
returned as the function result, rather than passed as a
pointer to argument.

This ensures on error inode_table is set to NULL, which
removes the spurious double free later.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>