unsquashfs: Non-fatal error hardening part 1
If unsquashfs fails to read a directory from the
Squashfs filesystem treat it as a fatal error.

Originally this was treated as a non-fatal error,
and the sub-directory was skipped.  But, due to
current greater concerns about corruption, any failure
to read the Squashfs filesystem should be now treated as
a fatal error.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>