unsquashfs: Non-fatal error hardening part 2
If Unsquashfs fails to create a directory, file, or
hard-link a file, now treat this a fatal error.

This is likely due to out of space on the output filesystem,
or some other fatal I/O errors.  This should always have been
treated as fatal.

Failure to create other types (not directory or file, e.g.
character device or fifo) is still treated as non-fatal
(but the exit code will reflect failure), as the destination
filesystem may not support that type.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>