unsquashfs: Non-fatal error hardening part 3
If the writer thread fails to write a block to the output
filesystem, treat this as a fatal error.

This is likely due to out of space on the output filesystem,
or some other fatal I/O errors.  This should always have been
treated as fatal.

Additionally, if the writer receives an errored buffer,
which means the Squashfs filesystem could not be read
or decompressed, then also now treat this as a fatal error.

Originally this was treated as a non-fatal error,
and the file was skipped.  But, due to current greater concerns
about corruption, any failure to read the Squashfs filesystem
should be now treated as a fatal error.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>