unsquashfs: Fix previous patch to use octal and fix BUG in it
The previous patch improves sticky bit handling in
Unsquashfs set_attributes, by changing the code to
attempt to set the sticky-bit even as a user-process,
and only then on failure, try to drop the sticky-bit.

Unfortunately, the patch has a nasty bug, and comes with another
patch dependency that gratuitously replaces perfectly good octal
values with a variable, and which is part of the cause of the bug!

i.e.

07000 is replaced by root_mask, defined by

mode_t root_mask = S_ISUID | S_ISGID | S_ISVTX;

When it comes to permissions *I* like to see the bits, and
I disagree doing the above makes things clearer.  In fact
I think it obfusticates things.  So I dropped the
dependency patch which changed these values.

The really "funny" thing is that there is a bug in the
patch which becomes obvious when octal values are used.
But it is not so obvious when the variable root_mask is
used instead.

The BUG:

The patch changes root_mask to be S_ISUID | S_ISGID.

So when you have a user-process the sticky-bit is not removed in
the line

mode &= ~(root_mask);

It then executes these lines:

if((set_mode || (mode & root_mask)) && chmod(pathname, (mode_t) mode) ==
-1) {
	if (root_process || errno != EPERM || !(mode & S_ISVTX) ||
		chmod(pathname, (mode_t) (mode & ~S_ISVTX)) == -1)

See the bug?  You probably will do that now I've pointed it out.

The root_mask doesn't have the sticky-bit set, and so it will
always evaluate to 0 for a user-process (as the same mask was
used to remove the bits previously).

So the code will not try to set the sticky-bit for a user-process.
In fact it won't try to set the sticky-bit for a root process either
if bits 06000 are not set (both cases ignoring set_mode).
The exact opposite of what the patch is supposed to do!

If octal values had been used, it would have been much more obvious
there are bugs!

mode &= ~06000;

if((set_mode || (mode & 06000)) && chmod(pathname, (mode_t) mode) == -1)
{
	if (root_process || errno != EPERM || !(mode & 01000) ||
		chmod(pathname, (mode_t) (mode & ~01000)) == -1)

The two operations with 06000 are obviously the exact opposite, and
so the first chmod will never execute (ignoring set_mode).

Additionally in the second if, (mode & 01000) is testing a bit
that was previously not checked (mode & 06000), which further implies
there is something wrong with the logic.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>