unsquashfs: xattr: Make failure to write system xattrs non-fatal again
Trying to unsquash a filesystem which has system xattrs as
non-root causes Unsquashfs to abort.

This is one of those errors I made fatal because it results in
loss of xattrs.

But in testing it is proving to be a PITA.   All of test
filesystems like that is is now causing Unsquashfs to
abort.

If that's annoying to me in testing, it is probably not
a good idea.

So make this abort only if -strict-errors is set.

Unsquashfs will return an error exit code without making
this fatal, and that should be good enough.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>