Unsquashfs: Add option to display time in UTC
Unsquashfs has always displayed time in the local time zone
(which can be over-ridden by the TZ environment variable).

While this is entirely correct, and is how ls -la behaves,
it can be a bit annoying when comparing against the
UTC time taken by -mkfs-time and -all-time options
of Mksquashfs.

It also is a useful additional option in its own right.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>