mksquashfs: pad with 0xff to accommodate FLASH
This patch changes mksquashfs to pad generated images with
a 0xff pattern rather than 0x00. The reason being that when
a image is written to SLC NOR/NAND flash, the individual cell
that will hold the padding needs to be "programmed" to produce
the "0" when read. This will reduce a tiny bit of wear and
tear in this situations and should be harmless to all other
contexts in which squashfs is in use.

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>