unsquashfs: attempt chmod and chown when running as non-root with strict-errors
We interpret 'strict-errors' as 'fail if you cannot correctly reproduce
the original filesystem'. Without this change, unsquashfs does not even
attempt to chown files to their original owners when not running as
root, and failures to set the sticky bit via chmod are silently ignored.

Now we will attempt to chown, and fail, if strict-errors is set, and
failure to set the sticky bit will be treated as fatal.