Unsquashfs: Add new -max-depth <levels> option
Add a new option, similar to find -maxdepth, that limits
the unsquashing or listing of files, to <levels> directories.
Or only descend <levels> directories when unsquashing or
listing.

This is an option I have sometimes wanted with Unsquashfs to
limit the output, but as so often happens it is easier
to work around it when that happens, rather than go to the
trouble of adding it, and so it has remained a to-do feature.

A recent request suggests it is an option that other people
would find useful, and so here it is.   Better late than never.

It does what it says, a <levels> of 0 produces only the root
directory, a <levels> of 1 produces the entries in the
top-level directory and so on.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>