Unsquashfs: improve performance in inode table reading
I have received a report that when reading very large
filesystems, a large amount of time is spent reallocing
the inode table, due to the fact the realloc size used is
SQUASHFS_METADATA_SIZE (8 Kbytes).

This commit improves performance by basing the allocation
size on the compressed size of the inode table.

For different compression ratios, this will significantly
reduce allocs.

With a 50% compression ratio, this should require 2 alloc calls
With a 25% compression ratio, this should require 4 alloc calls
With a 12.5% compression ratio, this should require 8 alloc calls

Reported-by: da <da@babadook>
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>