Unsquashfs: improve performance in directory table reading
When reading very large filesystems, a large amount of time is
spent reallocing the directory table, due to the fact the realloc
size used is SQUASHFS_METADATA_SIZE (8 Kbytes).

This commit improves performance by basing the allocation
size on the compressed size of the directory table.

For different compression ratios, this will significantly
reduce allocs.

With a 50% compression ratio, this should require 2 alloc calls
With a 25% compression ratio, this should require 4 alloc calls
With a 12.5% compression ratio, this should require 8 alloc calls

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>