Unsquashfs: add option to follow symlinks in extract files
An issue on Github (https://github.com/plougher/squashfs-tools/issues/77)
raises the problem that symlinks in extract files are
not followed.  So if the extract file pathname ends with a
symlink (leaf component) it is extracted and left as a
dangling symlink, unless the real file it links to is
also specified as an extract file.

Additionally, if the extract file pathname traverses
symlinks while walking the pathname (i.e. non leaf
symlinks), the extraction will stop at this point.  In
this case the extract target (leaf component) will not
be extracted.  Here the problem isn't a dangling
symlink, but rather no symlink and nothing else
in the pathname followed.

One way of solving this is by dereferencing the symlink(s),
and replacing them with what they actually link to, in a
similar way to "cp -L".

But this is dangerous, and can cause Unsquashfs to produce
output which does not match the input Squashfs filesystem,
which is something Unsquashfs should never do.

The reason for this is whereas "cp -L" is derefencing
the *input*, Unsquashfs is dereferencing the *output*.

If a filesystem has a real filename, say

	a/b/c/hello_world

and two symlinks

a/symlink1 ---> b/c/hello_world
a/symlink2 ---> b/c

There are multiple different ways (or paths) to the *single*
hello_world file.

a/b/c/hello_world
a/symlink1
a/symlink2/hello_world

If Unsquashfs was given all three paths as extract files, and
Unsquashfs dereferenced them on output, you will get
*three* copies of the hello_world file, and two copies of
directory "c".

Superficially the output may look the same, but, it may
not work the same, and obviously edits to the one hello_world
file will not get reflected in the other copies.

Any option that can be accidently or maliciously used to produce
such an output is too dangerous to be added.

This commit solves the problem in an equivalent way, but which
does not alter the output, and so it is a completely safe option.

An extract file with symlinks is ultimately unresolvable because
it is not the canonical pathname to the file (that is the one
without symlinks), and without that the real filename and path
will not be extracted.

Additionally, if an extract file with symlinks references other
symlinks, then it is also unresolvable because these symlinks
will also not appear in the output.

Therefore to make an extract file with symlinks resolvable you
have to find the canonical pathname to the file, and add it
as an extract file.  Additionally any symlinks found while
resolving it also have to be added as an extract file.  Once
done the extract file will be resolvable.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>