Unsquashfs: add -no-exit-code option for non fatal errors
An issue on GitHub (https://github.com/plougher/squashfs-tools/issues/94)
raises the issue that errors that used to be treated
as non-fatal in earlier versions of Unsquashfs are now
treated as fatal, and that they set the exit code to nonzero.

Additionally, -ignore-errors can be used to make
Unsquashfs behave more like earlier versions, but, the
exit code is still set to nonzero.

This change in behaviour is deliberate, and it is in
response to requests to harden the error handling
of Unsquashfs, including always setting the exit code
(including non fatal errors).

But it is clear some people still want the original
behaviour where non-fatal errors did not set the
exit code.

This commit adds the -no-exit-code option which
will disable setting of the exit code when
non fatal errors occur.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>