Unsquashfs: add type field to extract pathname structure
Unsquashfs combines all the extract files specified
on the commandline into a hierarchy of directories
and leaf components (the actual extract target).

Previously the difference between a link component and
a leaf component was indicated by whether the component
had a sub-directory or was NULL.

Add explicit types to make the code clearer, and
to be able to handle more types.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>