Unsquashfs: add code to handle -exclude files
This commit adds code to create a walkable hierarchy
of exclude files/directories.  It also adds
the code to match against exclude files in the
pre_scan and dir_scan functions, which do the
walking/output of the Squashfs filesystem
directory hierarchy.

If only extract files are specified, then
Unsquashfs behaviour is unchanged - only
the specific extract files are extracted
from the filesystem.

If only exclude files ares specified, then
Unsquashfs will extract everything, but, every
file/directory which matches the exclude
files will be excluded.

If both extract and exclude files are specified,
then Unsquashfs will match the extract files
to produce the sub-set of files to be
extracted, and then the exclude files will be
applied to that sub-set, to produce the
final set of files.

This commit does not add any exclude options
code.  This will be added in the next commits.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>