Unsquashfs: add -exclude-file option
This takes a file of names to be excluded from the
output, similar to the extract option -ef.

Due to the now large amount of options in Unsquashfs,
meaningful "short" versions of options  like -e and -ef
are becoming impossible to achieve.

This adds -excf and -exc as alternative short versions.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>