Mksquashfs: Add I/O and CPU throttling option
An issue on GitHub (https://github.com/plougher/squashfs-tools/issues/97)
raises the problem that Mksquashfs can cause some computers
to over-heat, even with the -processors option set to 1.

The reason for this is because Mksquashfs is highly parallelised, and
spawns multiple compression threads (both block and fragment),
by default one each of those compression threads are spawned per core.
In addition it also has separate reader and writer threads and
large caches to ensure I/O through-put is maximised and
performed in parallel with the compression threads (called
read-ahead and write-behind).

This means even when the "-processors 1" option is set, there
is still a lot of parallel processing being performed, because
the minmimum set of threads is a single block compression thread,
a single fragment compression thread, and both a reader and writer
thread.  Unfortunately, these threads can easily still max out a
single core, so that it is never idle.  On a multi-core machine the
"-processors 1" option usually produces a CPU load of around 130-180%.

So there has been for quite a while a need for a way to
reduce the I/O and CPU use of Mksquashfs to genuinely below
1 core, and if possible to significantly below 1 core.  This isn't
only needed for computers that will over-heat, it is also
sometimes convienient to reduce the I/O consumption of
Mksquashfs when it is being run in the background with other
tasks.  With the power of modern cores, even "-processors 1" can
produce high I/O load.

The design of Mksquashfs means any I/O (and thence CPU) limiting
measures should be in the reader thread.  This is the source
of the data to the other threads, and if it is reduced or
throttled in the reader thread, then the other threads will
automatically idle.

But the throttling measures in the reader thread should be
non-invasive, and cause no reduction in performance when no
throttling is in effect.

The solution adopted is to add a 10 second timer interrupt
(SIGALRM signal), and then have the interrupt handler
idle for a percentage of the 10 second cycle.  By increasing
or decreasing that idle percentage you can reduce or increase
the throughput of the reader thread, with each percentage point
(between 0 and 100) added to the sleep time reducing the
time the reader thread runs by 0.1 seconds.

Two complementary options have been added to control this
behaviour.

-throttle <percentage> has the semantics of throttle
the reader by X%.  The higher the percentage, the
greater the throughput of the reader thread is reduced.

-limit <percentage> has the opposite sense of limit
the output to X%.  The higher the percentage, the
greater the throughput of the reader thread.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>