Mksquashfs: add -no-hardlinks option
Normally Mksquashfs will detect hardlinks (multiple files with the
same inode) and hardlink them in the Squashfs image.

But often hardlinks are used in the original filesystem
to save space, when files are discovered to be duplicate.
In this special case the only reason the files are
hardlinked is to save space, and where the filesystem
doesn't handle duplicate files (different inode, same
data).

Squashfs does handle duplicate files, and so add
an option to ignore hardlinks and instead
store them as duplicates.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>