Pseudo: Add support for Pseudo file hardlinks
Pseudo files had the ability to create block devices,
character devices, directories, regular files and
symlinks.  The option that was missing was the ability
to create hardlinks.

This commit adds the ability to hardlink to a file, that
is create a reference to another file.

The syntax is

"filename l pathname"

which will create a file named "filename" hardlinked to the
source "pathname".  As you would expect, it works in the
same way as doing a "ln pathname filename" in the Squashfs
source directories.

Pathname is absolute if prefixed with /, otherwise it is
treated as relative to the current working directory.  Any
file can be hardlinked, and it can be inside or outside the
Mksquashfs source directories.

If the above created hardlink is the only reference to the
file in the Mksquashfs source directories, it will be created
and it will have an nlink count of 1.

If there are multiple references to the file within the
Mksquashfs source directories, then a single file is created,
and the nlink count is incremented for each reference.

The Pseudo file hardlink option can be used for any
file type that can be hardlinked, which means
block devices, character devices, sockets, regular
files and symlinks, but, not directories.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>