Pseudo: Restructure pseudo structures part 2
Make the stat structures (struct stat and struct pseudo_stat)
a union of pointers.

This allows that pointer to be used for both the Link option
and the other options.

This allows the "char *" union of pointers to directly
store the linkname for the Link option.

This allows the "struct pseudo_dev_link" structure to
be removed.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>