Pseudo: move creation of timestamp and ino to pseudo parsing
Move the creation of the timestamp and ino to creation of the
"struct pseudo_dev" when the pseudo definitions are parsed.

The difference is these become unique to the "struct pseudo_dev"
structure, rather than at the moment being completely unique,
where every reference even to the same "struct pseudo_dev"
will produce different results.

In otherwords now every reference to the same "struct pseudo_dev"
structure will have the same values.  This is essential to
detect hardlinks to the pseudo definition files.

The recently added Pseudo Link option cannot be used
to create hardlinks to files created by a Pseudo definition
as they (obviously) don't exist in the source filesystem.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>