Pseudo: Add support for Hardlinking to Pseudo definitions
The recently added Pseudo file hardlink option works
for "real" files which exist in the filesystem (anywhere
under "/").  These can be absolute (from /) or relative
to the current working directory.

But it can't be used to create a Pseudo hardlink to a
file itself created by a Pseudo definition, because it
doesn't exist in the source filesystem.

One way to add this functionality would be to see if it exists
as a "real" file in the source filesystem, and then fall
back to see if it has been created by a Pseudo definition.
Here we will be looking at the output filesystem not the
source filesystem.  Obviously previously we don't want
to restrict the linking to what is already in the output
filesystem.

The problem with this approach is we have two different
namespaces in operation.  When you mix namespaces you can
get lots of mix-ups where things match on the wrong thing etc.
In general it is better to avoid this.

So this commit adds an explicit add hardlink to a
Pesudo file option.  This is distinquished by an
uppercase L, i.e.

a/b/c L 1/23

Will create a hardlink to the Pseudo created file "1/2/3"
named "a/b/c".

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>