Pseudo: Add support for extended definitions with timestamp
Add support for extended definition with timestamp.

These are similar to existing definitions, but with an
uppercase type.

i.e.

file D time mode uid gid

Will create a directory with timestamp time, and with
mode uid and gid as before.

Currently, time must be an unisgned decimal integer, which
represents the seconds since the epoch of 1970-01-01 01:00 UTC.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>