Unsquashfs: Add code to dynamically read inode table
Originally when Unsquashfs was written it decompressed
the entire filesystem, and wrote it out to disk.
As such it made sense to read and decompress the
entire inode table at start-up.

Since then new options have been added to Unsquashfs which
can significantly limit the amount of the filesystem
written to disk, e.g. extract files, and more recently
exclude files and depth limiting (-max option).

So it no longer makes sense to read and decompress
the entire inode table at start-up, when much of it
may never be referenced.  It just adds unnecessary
overhead.

This is the first in a series of patches, which removes
the start-up reading of the tables, and moves over
to dynamic reading on demand.

This patch adds the necessary metadata reading functions
to Unsquashfs, plus the necessary locking in read_fs_bytes()
because there now will be simultaneous reads to the
filesystem image by the reader thread and the main
thread.

Later patches will update the separate Unsquashfs
implementations for 4.x, 3.x, 2.x and 1.x.

The patches are written to be incremental, keeping the tools
functional until everything is converted, and the
original code can be removed.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>