Unsquash-3: update to dynamically read inode table
Originally when Unsquashfs was written it decompressed
the entire filesystem, and wrote it out to disk.
As such it made sense to read and decompress the
entire inode table at start-up.

Since then new options have been added to Unsquashfs which
can significantly limit the amount of the filesystem
written to disk, e.g. extract files, and more recently
exclude files and depth limiting (-max option).

So it no longer makes sense to read and decompress
the entire inode table at start-up, when much of it
may never be referenced.  It just adds unnecessary
overhead.

This is the third patch in the series, which
updates the Unsquashfs 3.x implemention to
dynamically read the inode table.

Note: the buffer list allocated in read_inode() is
leaked (not freed).  This code is a temporary workaround
to keep functionality working before read_block_list()
is updated in later patches (to do it now would break
the unconverted 2.x and 1.x implementations).

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>