Unsquash-3: update to dynamically read directory table
Originally when Unsquashfs was written it decompressed
the entire filesystem, and wrote it out to disk.
As such it made sense to read and decompress the
entire directory table at start-up.

Since then new options have been added to Unsquashfs which
can significantly limit the amount of the filesystem
written to disk, e.g. extract files, and more recently
exclude files and depth limiting (-max option).

So it no longer makes sense to read and decompress
the entire directory table at start-up, when much of it
may never be referenced.  It just adds unnecessary
overhead.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>