Unsquashfs: Add missing check for Symlink type
If extract_matches() returned a NULL sub-directory and the
inode type is not a directory, we extract the leaf file.

If, however, extract_matches() has returned a sub-directory
and this isn't a directory, then there is mis-match
(think where someone has done "name/wildcard/name, and
wildcard has matched on a non directory).  In this
case the match is ignored.

However, the special case of a symlink type is not checked
here, and Unsquashfs will therefore not free the symlink
storage.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>