Unsquashfs: -cat option, replace stack_pathname() with another implementation
The stack in cat_scan() keeps track of the parent directories
visited and the current depth of traversal while walking the supplied
pathname.

The stack is essential to allow  ".", ".." and in the future
symbolic links in the pathname to be correctly walked.

At any one time the stack represents the "real" current
position and depth of traversal in the directory hierarchy
following walking of the above elements (., .., and symbolic
links), or in otherwords, we can go up the hierarchy following
".." and symbolic links, as well as down.

In effect the stack represents the result of following
the above while walking the pathname.  This produces
a canonicalised pathname which will be different from
the supplied pathname if it involved traversing ".", ".."
and symbolic links.

Why is this important? It is important because stack_pathname()
will produce the canonicalised pathname, rather then the
actual supplied pathname.  This could be confusing when
error messages give the canonicalised pathname, rather
than the actual pathname supplied.

For example, the supplied pathname

squashfs-tools/././squashfs-tools/PATCHES/..

is invalid, because it resolves to a directory (squashfs-tools),
but, the canonicalised pathname which would be displayed
in the error message would be

squashfs-tools/squashfs-tools

Which isn't the pathname supplied by the user.  Here the
user should be able to work out how the pathname relates
to the supplied pathname, but, they shouldn't have to do it.

But things will get a lot more confusing when symbolic
links are supported.

Imagine if we had a symbolic link called "x" with contents "../../dir",
and this was used in a pathname, e.g.

a/b/c/x/dir2

Say "dir2" is a directory, which will cause an error.  This would
produce the following pathname in the error message

a/dir/dir2 is a directory.

which doesn't look very much like the supplied pathname.

So unlike -follow-symlinks where the canonicalised pathname
returned by stack_pathname() is exactly what we want, we have
to move to a different way of tracking the pathname descent
which preserves ".", "..", and symbolic links.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>